+++ Sources/Vapor/Session/Request+Session.swift
@@ -1,0 +1,11 @@
+extension Request {
+    /// Server stored information related from session cookie.
+    public var session: Session? {
+        get {
+            return storage["session"] as? Session
+        }
+        set(session) {
+            storage["session"] = session
+        }
+    }
+}
