+++ Tests/Vapor/RequestTests.swift
@@ -41,7 +41,7 @@ class RequestTests: XCTestCase {
     func testParse() {
         let string = "value=123"
 
-        let data = Request.parseFormURLEncoded(string.data)
+        let data = FormURLEncoded.parse(string.data)
         XCTAssert(data["value"]?.int == 123, "Request did not parse correctly")
     }
 
@@ -55,7 +55,7 @@ class RequestTests: XCTestCase {
         body += "--" + boundary + "\r\n"
 
         let parsedBoundary = try! Multipart.parseBoundary(contentType: "multipart/form-data; charset=utf-8; boundary=\(boundary)")
-        let data = Multipart.parseMultipartForm(body.data, boundary: parsedBoundary)
+        let data = Multipart.parse(body.data, boundary: parsedBoundary)
 
         XCTAssert(data["value"]?.int == 123, "Request did not parse correctly")
     }
@@ -71,7 +71,7 @@ class RequestTests: XCTestCase {
         body += "--" + boundary + "\r\n"
 
         let parsedBoundary = try! Multipart.parseBoundary(contentType: "multipart/form-data; charset=utf-8; boundary=\(boundary)")
-        let data = Multipart.parseMultipartForm(body.data, boundary: parsedBoundary)
+        let data = Multipart.parse(body.data, boundary: parsedBoundary)
 
         XCTAssert(data["value"]?.file?.data == "123".data, "Request did not parse correctly")
     }
@@ -79,7 +79,7 @@ class RequestTests: XCTestCase {
     func testFormURLEncoded() {
         let body = "first=value&arr[]=foo+bar&arr[]=baz"
 
-        let data = Request.parseFormURLEncoded(body.data)
+        let data = FormURLEncoded.parse(body.data)
 
         XCTAssert(data["first"]?.string == "value", "Request key first did not parse correctly")
         XCTAssert(data["arr", 0]?.string == "foo bar", "Request key arr did not parse correctly")
@@ -89,7 +89,7 @@ class RequestTests: XCTestCase {
     func testFormURLEncodedEdge() {
         let body = "singleKeyArray[]=value&implicitArray=1&implicitArray=2"
 
-        let data = Request.parseFormURLEncoded(body.data)
+        let data = FormURLEncoded.parse(body.data)
 
         XCTAssert(data["singleKeyArray", 0]?.string == "value", "singleKeyArray did not parse correctly")
         XCTAssert(data["implicitArray", 0]?.string == "1", "implicitArray did not parse correctly")
